api/models/video.py
from django.conf import settings
from django.db import models
from django.utils.translation import ugettext_lazy as _
from model_utils.models import TimeStampedModel


class Video(TimeStampedModel):

    description = models.TextField()

    url = models.URLField()

    thumbnail = models.URLField()

    # relations

    author = models.ForeignKey(settings.AUTH_USER_MODEL,
                               null=True, on_delete=models.SET_NULL,
                               related_name="created_videos")
