requirements.txt
Django==2.0.3
django-cors-headers==2.2.0
django-filter==1.1.0
django-model-utils==3.1.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
drf-writable-nested==0.4.1
PyJWT==1.6.1
pytz==2018.3
